Made MkSocket as a standalone class
